+++ src/tests/EmptyPostDetail.test.tsx
@@ -1,0 +1,9 @@
+import { test } from "vitest";
+import { render } from "@testing-library/react";
+
+import EmptyPostDetail from "../components/EmptyPostDetail";
+
+test("EmptyPostDetail - Render test", async () => {
+  const component = render(<EmptyPostDetail />);
+  await component.findByText("No post selected");
+});