+++ src/tests/utils/__mocks__.ts
@@ -1,0 +1,12 @@
+export const post: IPost = {
+  id: "t3_t31tek",
+  title:
+    "Ukrainian President Zelenskyy having a meal with his soldiers. A genuine leader for the people.",
+  author: "-Zepphyrrez-",
+  created: "hace 13 horas",
+  numComments: 3023,
+  thumbnail:
+    "https://b.thumbs.redditmedia.com/WUyR2jZ6MHD1bxAHHSqSbZaVQLRmoaBxrSqm9pmLbfI.jpg",
+  image: "https://i.redd.it/khzw5r5xugk81.jpg",
+  opened: false,
+};