+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.1.5-dev
 
+- [NEW] Render tests added for all components @christianmotta
 - [NEW] Adding testing dependencies and some settings @christianmotta
 
 ### v0.1.4